work on admin designs
Seo Url, Seo Regex, Banners

+++ admin/controller/design/seo_url.php
@@ -1,10 +1,12 @@
 <?php
 
-class ControllerDesignSeoUrl extends PT_Controller {
+class ControllerDesignSeoUrl extends PT_Controller
+{
 
     private $error = array();
 
-    public function index() {
+    public function index()
+    {
         $this->load->language('design/seo_url');
 
         $this->document->setTitle($this->language->get('heading_title'));
@@ -14,109 +16,109 @@ class ControllerDesignSeoUrl extends PT_Controller {
         $this->getList();
     }
 
-    public function add() {
+    public function add()
+    {
         $this->load->language('design/seo_url');
 
         $this->document->setTitle($this->language->get('heading_title'));
 
         $this->load->model('design/seo_url');
 
-        if (($this->request->server['REQUEST_METHOD'] == 'POST')) {
-          
+        if (($this->request->server['REQUEST_METHOD'] == 'POST') && $this->validateForm()) {
+
             $this->model_design_seo_url->addSeoUrl($this->request->post);
 
             $this->session->data['success'] = $this->language->get('text_success');
 
-            $this->response->redirect($this->url->link('design/seo_url', 'user_token=' . $this->session->data['user_token']));
+            $this->response->redirect($this->url->link('design/seo_url', 'member_token=' . $this->session->data['member_token']));
         }
 
         $this->getForm();
     }
 
-    public function edit() {
+    public function edit()
+    {
         $this->load->language('design/seo_url');
 
         $this->document->setTitle($this->language->get('heading_title'));
 
         $this->load->model('design/seo_url');
 
-        if (($this->request->server['REQUEST_METHOD'] == 'POST')) {
+        if (($this->request->server['REQUEST_METHOD'] == 'POST') && $this->validateForm()) {
             $this->model_design_seo_url->editSeoUrl($this->request->get['seo_url_id'], $this->request->post);
 
             $this->session->data['success'] = $this->language->get('text_success');
 
-            $this->response->redirect($this->url->link('design/seo_url', 'user_token=' . $this->session->data['user_token']));
+            $this->response->redirect($this->url->link('design/seo_url', 'member_token=' . $this->session->data['member_token']));
         }
 
         $this->getForm();
     }
 
-    public function delete() {
+    public function delete()
+    {
         $this->load->language('design/seo_url');
 
         $this->document->setTitle($this->language->get('heading_title'));
 
         $this->load->model('design/seo_url');
 
-        if (isset($this->request->post['selected'])) {
+        if (isset($this->request->post['selected']) && $this->validateDelete()) {
             foreach ($this->request->post['selected'] as $seo_url_id) {
                 $this->model_design_seo_url->deleteSeoUrl($seo_url_id);
             }
 
             $this->session->data['success'] = $this->language->get('text_success');
 
-            $this->response->redirect($this->url->link('design/seo_url', 'user_token=' . $this->session->data['user_token']));
+            $this->response->redirect($this->url->link('design/seo_url', 'member_token=' . $this->session->data['member_token']));
         }
 
         $this->getList();
     }
 
-    protected function getList() {
-        $this->document->addStyle("view/dist/plugins/DataTables/DataTables-1.10.18/css/dataTables.bootstrap4.min.css");
-        $this->document->addStyle("view/dist/plugins/DataTables/Buttons-1.5.6/css/buttons.bootstrap4.min.css");
-        $this->document->addStyle("view/dist/plugins/DataTables/FixedHeader-3.1.4/css/fixedHeader.bootstrap4.min.css");
-        $this->document->addStyle("view/dist/plugins/DataTables/Responsive-2.2.2/css/responsive.bootstrap4.min.css");
-        $this->document->addScript("view/dist/plugins/DataTables/DataTables-1.10.18/js/jquery.dataTables.min.js");
-        $this->document->addScript("view/dist/plugins/DataTables/DataTables-1.10.18/js/dataTables.bootstrap4.min.js");
-        $this->document->addScript("view/dist/plugins/DataTables/Buttons-1.5.6/js/dataTables.buttons.min.js");
-        $this->document->addScript("view/dist/plugins/DataTables/Buttons-1.5.6/js/buttons.bootstrap4.min.js");
-        $this->document->addScript("view/dist/plugins/DataTables/JSZip-2.5.0/jszip.min.js");
-        $this->document->addScript("view/dist/plugins/DataTables/pdfmake-0.1.36/pdfmake.min.js");
-        $this->document->addScript("view/dist/plugins/DataTables/pdfmake-0.1.36/vfs_fonts.js");
-        $this->document->addScript("view/dist/plugins/DataTables/Buttons-1.5.6/js/buttons.html5.min.js");
-        $this->document->addScript("view/dist/plugins/DataTables/Buttons-1.5.6/js/buttons.print.min.js");
-        $this->document->addScript("view/dist/plugins/DataTables/Buttons-1.5.6/js/buttons.colVis.min.js");
-        $this->document->addScript("view/dist/plugins/DataTables/FixedHeader-3.1.4/js/dataTables.fixedHeader.min.js");
-        $this->document->addScript("view/dist/plugins/DataTables/FixedHeader-3.1.4/js/fixedHeader.bootstrap4.min.js");
-        $this->document->addScript("view/dist/plugins/DataTables/Responsive-2.2.2/js/dataTables.responsive.min.js");
-        $this->document->addScript("view/dist/plugins/DataTables/Responsive-2.2.2/js/responsive.bootstrap4.min.js");
+    protected function getList()
+    {
+        $this->document->addScript("view/dist/js/jquery.dataTables.min.js");
+        $this->document->addScript("view/dist/js/jquery.dataTables.bootstrap.min.js");
+        $this->document->addScript("view/dist/js/dataTables.buttons.min.js");
+        $this->document->addScript("view/dist/js/buttons.flash.min.js");
+        $this->document->addScript("view/dist/js/buttons.html5.min.js");
+        $this->document->addScript("view/dist/js/buttons.print.min.js");
+        $this->document->addScript("view/dist/js/buttons.colVis.min.js");
+        $this->document->addScript("view/dist/js/dataTables.select.min.js");
 
         $data['breadcrumbs'] = array();
 
         $data['breadcrumbs'][] = array(
             'text' => $this->language->get('text_home'),
-            'href' => $this->url->link('common/dashboard', 'user_token=' . $this->session->data['user_token'])
+            'href' => $this->url->link('common/dashboard', 'member_token=' . $this->session->data['member_token'])
         );
 
         $data['breadcrumbs'][] = array(
             'text' => $this->language->get('heading_title'),
-            'href' => $this->url->link('design/seo_url', 'user_token=' . $this->session->data['user_token'])
+            'href' => $this->url->link('design/seo_url', 'member_token=' . $this->session->data['member_token'])
         );
 
-        $data['add'] = $this->url->link('design/seo_url/add', 'user_token=' . $this->session->data['user_token']);
-        $data['delete'] = $this->url->link('design/seo_url/delete', 'user_token=' . $this->session->data['user_token']);
+        $data['add'] = $this->url->link('design/seo_url/add', 'member_token=' . $this->session->data['member_token']);
+        $data['delete'] = $this->url->link('design/seo_url/delete', 'member_token=' . $this->session->data['member_token']);
 
         $data['seo_urls'] = array();
 
-        $results = $this->model_design_seo_url->getSeoUrls();
-        
+        $filter_data = array(
+            'sort'  => 'query',
+            'order' => 'ASC'
+        );
+
+        $results = $this->model_design_seo_url->getSeoUrls($filter_data);
+
         foreach ($results as $result) {
             $data['seo_urls'][] = array(
-                'seo_url_id' => $result['seo_url_id'],
-                'query' => $result['query'],
-                'keyword' => $result['keyword'],
-                'edit' => $this->url->link('design/seo_url/edit', 'user_token=' . $this->session->data['user_token'] . '&seo_url_id=' . $result['seo_url_id']),
-                'delete' => $this->url->link('design/seo_url/delete', 'user_token=' . $this->session->data['user_token'] . '&seo_url_id=' . $result['seo_url_id'])
+                'seo_url_id'    => $result['seo_url_id'],
+                'query'         => $result['query'],
+                'keyword'       => $result['keyword'],
+                'language'      => $result['language'],
+                'edit'          => $this->url->link('design/seo_url/edit', 'member_token=' . $this->session->data['member_token'] . '&seo_url_id=' . $result['seo_url_id']),
+                'delete'        => $this->url->link('design/seo_url/delete', 'member_token=' . $this->session->data['member_token'] . '&seo_url_id=' . $result['seo_url_id'])
             );
         }
 
@@ -147,7 +149,8 @@ class ControllerDesignSeoUrl extends PT_Controller {
         $this->response->setOutput($this->load->view('design/seo_url_list', $data));
     }
 
-    protected function getForm() {
+    protected function getForm()
+    {
         $this->document->addStyle("view/dist/plugins/iCheck/all.css");
         $this->document->addScript("view/dist/plugins/ckeditor/ckeditor.js");
         $this->document->addScript("view/dist/plugins/ckeditor/adapters/jquery.js");
@@ -183,35 +186,35 @@ class ControllerDesignSeoUrl extends PT_Controller {
 
         $data['breadcrumbs'][] = array(
             'text' => $this->language->get('text_home'),
-            'href' => $this->url->link('common/dashboard', 'user_token=' . $this->session->data['user_token'])
+            'href' => $this->url->link('common/dashboard', 'member_token=' . $this->session->data['member_token'])
         );
 
         $data['breadcrumbs'][] = array(
             'text' => $this->language->get('heading_title'),
-            'href' => $this->url->link('design/seo_url', 'user_token=' . $this->session->data['user_token'])
+            'href' => $this->url->link('design/seo_url', 'member_token=' . $this->session->data['member_token'])
         );
 
         if (!isset($this->request->get['seo_url_id'])) {
-            $data['action'] = $this->url->link('design/seo_url/add', 'user_token=' . $this->session->data['user_token']);
+            $data['action'] = $this->url->link('design/seo_url/add', 'member_token=' . $this->session->data['member_token']);
             $data['breadcrumbs'][] = array(
                 'text' => $this->language->get('text_add'),
-                'href' => $this->url->link('design/seo_url/add', 'user_token=' . $this->session->data['user_token'])
+                'href' => $this->url->link('design/seo_url/add', 'member_token=' . $this->session->data['member_token'])
             );
         } else {
-            $data['action'] = $this->url->link('design/seo_url/edit', 'user_token=' . $this->session->data['user_token'] . '&seo_url_id=' . $this->request->get['seo_url_id']);
+            $data['action'] = $this->url->link('design/seo_url/edit', 'member_token=' . $this->session->data['member_token'] . '&seo_url_id=' . $this->request->get['seo_url_id']);
             $data['breadcrumbs'][] = array(
                 'text' => $this->language->get('text_edit'),
-                'href' => $this->url->link('design/seo_url/edit', 'user_token=' . $this->session->data['user_token'] . '&seo_url_id=' . $this->request->get['seo_url_id'])
+                'href' => $this->url->link('design/seo_url/edit', 'member_token=' . $this->session->data['member_token'] . '&seo_url_id=' . $this->request->get['seo_url_id'])
             );
         }
 
-        $data['cancel'] = $this->url->link('design/seo_url', 'user_token=' . $this->session->data['user_token']);
+        $data['cancel'] = $this->url->link('design/seo_url', 'member_token=' . $this->session->data['member_token']);
 
         if (isset($this->request->get['seo_url_id']) && ($this->request->server['REQUEST_METHOD'] != 'POST')) {
             $seo_url_info = $this->model_design_seo_url->getSeoUrl($this->request->get['seo_url_id']);
         }
-       
-        $data['user_token'] = $this->session->data['user_token'];
+
+        $data['member_token'] = $this->session->data['member_token'];
 
         if (isset($this->request->post['query'])) {
             $data['query'] = $this->request->post['query'];
@@ -248,7 +251,7 @@ class ControllerDesignSeoUrl extends PT_Controller {
         } else {
             $data['language_id'] = '';
         }
-       
+
         $data['header'] = $this->load->controller('common/header');
         $data['nav'] = $this->load->controller('common/nav');
         $data['footer'] = $this->load->controller('common/footer');
@@ -256,7 +259,8 @@ class ControllerDesignSeoUrl extends PT_Controller {
         $this->response->setOutput($this->load->view('design/seo_url_form', $data));
     }
 
-    protected function validateForm() {
+    protected function validateForm()
+    {
         if (!$this->member->hasPermission('modify', 'design/seo_url')) {
             $this->error['warning'] = $this->language->get('error_permission');
         }
@@ -276,7 +280,7 @@ class ControllerDesignSeoUrl extends PT_Controller {
         }
 
         if ($this->request->post['keyword']) {
-            if (preg_match('/[^a-zA-Z0-9_\-]+/', $this->request->post['keyword'])) {
+            if (preg_match('/[^a-zA-Z0-9_-]|[\p{Cyrillic}]+/u', $this->request->post['keyword'])) {
                 $this->error['keyword'] = $this->language->get('error_keyword');
             }
 
@@ -298,12 +302,12 @@ class ControllerDesignSeoUrl extends PT_Controller {
         return !$this->error;
     }
 
-    protected function validateDelete() {
+    protected function validateDelete()
+    {
         if (!$this->member->hasPermission('modify', 'design/seo_url')) {
             $this->error['warning'] = $this->language->get('error_permission');
         }
 
         return !$this->error;
     }
-
 }
